requirements.txt
gpt4all==2.8.2
langchain==0.3.7
google-generativeai==0.8.3
pydantic==2.9.2
python-dotenv==1.0.1
streamlit==1.40.1
langchain-core==0.3.15
langchain-google-genai==2.0.4
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langchain-chroma==0.1.4
langgraph==0.2.46

